chore: move dashboard and user management page folder
